Move bsd specifics into separate module
